# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DNDWORLDAPP API",
    "version": "0.1.0",
    "description": "Self-hosted worldbuilding and campaign API.\n\nTwo ways to authenticate: a `dwa_session` cookie for the browser, or\n`Authorization: Bearer <token>` for scripts and the MCP server.\n\nA token acts as its owner and inherits that user's role in each world.\nScopes and the optional world pin only ever *narrow* that — a token can\nnever grant more than the person holding it has.\n\nThings the caller may not see answer 404 rather than 403, so ids cannot\nbe probed to discover hidden content.\n\nThis document is generated from the Zod schemas in packages/schema."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
...
  "paths": {
...
    "/tokens": {
      "get": {
        "tags": [
          "tokens"
        ],
        "summary": "Your API tokens",
        "description": "Session only. A token cannot manage tokens.",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tokens": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TokenDto"
                      }
                    }
                  },
                  "required": [
                    "tokens"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tokens"
        ],
        "summary": "Mint a token; the secret is returned once and never again",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTokenInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "$ref": "#/components/schemas/TokenDto"
                    },
                    "secret": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token",
                    "secret"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in, or the token is unknown, revoked or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed — including a token missing a scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found. Also returned for things the caller may not see, so ids cannot be probed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TokenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "worldId": {
            "type": "string",
            "nullable": true
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "world:read",
                "world:write",
                "admin"
              ]
            }
          },
          "prefix": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer"
          },
          "lastUsedAt": {
            "type": "integer",
            "nullable": true
          },
          "expiresAt": {
            "type": "integer",
            "nullable": true
          },
          "revokedAt": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "worldId",
          "scopes",
          "prefix",
          "createdAt",
          "lastUsedAt",
          "expiresAt",
          "revokedAt"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "CreateTokenInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "worldId": {
            "type": "string",
            "nullable": true
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "world:read",
                "world:write",
                "admin"
              ]
            },
            "minItems": 1,
            "default": [
              "world:read"
            ]
          },
          "expiresInDays": {
            "type": "integer",
            "minimum": 1,
            "maximum": 3650,
            "nullable": true
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dwa_session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "A token from POST /tokens. Format: `dwa_` followed by 64 hex characters."
      }
    }
  }
}